000-apiflask_bootstrap/pyproject.toml
[tool.poetry]
name = "apiflask_bootstrap"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
APIFlask = "^1.3.1"
loguru = "^0.7.0"
peewee = "^3.16.2"
pycryptodome = "^3.18.0"
pymysql = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
